getFullTemplateByID
